requirements.txt
# 项目复杂度分析器依赖包
# 核心依赖
PyYAML>=6.0          # YAML配置文件支持
psutil>=5.8.0         # 系统性能监控

# 可选依赖（用于特定功能）
# requests>=2.25.0    # HTTP请求（如果需要远程分析）
# matplotlib>=3.3.0   # 图表生成（如果需要可视化报告）
# pandas>=1.3.0       # 数据分析（如果需要高级统计）
